+++ modules/azure_k8s_service/tests/test_azure_k8s_service.py
@@ -11,9 +11,11 @@ class TestAzureK8sServiceProcessor:
     def test_all_good(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "azure_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "azure_dummy_config.yaml"
             ),
             None,
         )
@@ -49,9 +51,11 @@ class TestAzureK8sServiceProcessor:
     def test_prehook(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "azure_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "azure_dummy_config.yaml"
             ),
             None,
         )
