+++ modules/cloudfront_distribution/tests/test_cloudfront_distribution.py
@@ -12,9 +12,11 @@ class TestAwsCloudfrontDstributionProcessor:
     def test_all_good(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config1.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config1.yaml"
             ),
             None,
         )
