+++ modules/gcp_k8s_service/tests/test_gcp_k8s_service.py
@@ -12,9 +12,11 @@ class TestGCPK8sServiceProcessor:
     def test_all_good(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "gcp_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "gcp_dummy_config.yaml"
             ),
             None,
         )
@@ -55,9 +57,11 @@ class TestGCPK8sServiceProcessor:
     def test_bad_postgres_permission(self):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "gcp_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "gcp_dummy_config.yaml"
             ),
             None,
         )
@@ -71,9 +75,11 @@ class TestGCPK8sServiceProcessor:
     def test_bad_redis_permission(self):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "gcp_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "gcp_dummy_config.yaml"
             ),
             None,
         )
@@ -87,9 +93,11 @@ class TestGCPK8sServiceProcessor:
     def test_bad_gcs_permission(self):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "gcp_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "gcp_dummy_config.yaml"
             ),
             None,
         )
