+++ modules/local_k8s_service/tests/test_local_k8s_service.py
@@ -11,9 +11,11 @@ class TestLocalK8sServiceProcessor:
     def test_all_good(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
@@ -58,9 +60,11 @@ class TestLocalK8sServiceProcessor:
     def test_handle_pg_link(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
@@ -86,9 +90,11 @@ class TestLocalK8sServiceProcessor:
     def test_handle_redis_link(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
@@ -114,9 +120,11 @@ class TestLocalK8sServiceProcessor:
     def test_handle_mysql_link(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
@@ -142,9 +150,11 @@ class TestLocalK8sServiceProcessor:
     def test_handle_mongodb_link(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
