+++ modules/tests/test_base.py
@@ -20,9 +20,11 @@ class TestBaseModuleProcessors:
     def test_validate_dns_delegated_false(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config_parent.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config_parent.yaml"
             ),
             None,
         )
@@ -40,9 +42,11 @@ class TestBaseModuleProcessors:
     def test_validate_dns_no_outputs(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config_parent.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config_parent.yaml"
             ),
             None,
         )
@@ -61,9 +65,11 @@ class TestBaseModuleProcessors:
     def test_validate_dns_no_name_servers(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config_parent.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config_parent.yaml"
             ),
             None,
         )
@@ -86,9 +92,11 @@ class TestBaseModuleProcessors:
     def test_validate_dns_mismatch(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config_parent.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config_parent.yaml"
             ),
             None,
         )
@@ -116,9 +124,11 @@ class TestBaseModuleProcessors:
     def test_validate_all_good(self, mocker: MockFixture) -> None:
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "dummy_config_parent.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "dummy_config_parent.yaml"
             ),
             None,
         )
