+++ modules/tests/test_linkerhelper.py
@@ -11,9 +11,11 @@ class TestLinkerHelper:
     def test_handle_link(self, mocker: MockFixture):
         layer = Layer.load_from_yaml(
             os.path.join(
-                os.path.dirname(os.path.dirname(__file__)),
-                "module_processors",
-                "local_dummy_config.yaml",
+                os.getcwd(),
+                "tests",
+                "fixtures",
+                "dummy_data",
+                "local_dummy_config.yaml"
             ),
             None,
         )
