+++ tests/fixtures/dummy_data/azure_dummy_config_parent.yaml
