+++ tests/fixtures/dummy_data/gcp_dummy_config.yaml
