+++ tests/fixtures/dummy_data/gcp_dummy_config_parent.yaml
