--- tests/module_processors/dummy_config_parent.yaml
@@ -1,17 +1,0 @@
-name: dummy-parent
-org_name: opta-tests
-providers:
-  aws:
-    region: us-east-1
-    account_id: 011111111111
-modules:
-  - type: aws-base
-  - type: aws-dns
-    domain: www.blah.com
-    upload_cert: true
-  - type: aws-eks
-    max_nodes: 6
-  - type: k8s-base
-  - type: datadog
-  - type: runx
-  - type: aws-ses
